+++ FalloutStrategy/FalloutEngine/j1Map.h
@@ -26,7 +26,7 @@ struct Properties
 	~Properties()
 	{
 
-		for (int i = 0; i < list.size(); i++)
+		for (size_t i = 0; i < list.size(); i++)
 		{
 			delete list[i];
 			list[i] = nullptr;
